Parceptron 0.003.0 Поправил расчет градиента на основе предыдущего градиента + тесты и перевел всё на double

+++ src/main/kotlin/Visualisation.kt
@@ -95,8 +95,8 @@ fun DrawScope.Layer(
     sectionCenter: Float,
     sectionHeight: Float,
     radius: Float,
-    neuronArray: Array<Float>,
-    weightsArray: Array<Array<Float>>,
+    neuronArray: Array<Double>,
+    weightsArray: Array<Array<Double>>,
     nextCenter: Float,
 ) {
     val heightOfNeuronArea = sectionHeight / neuronArray.count()
